fix: editor
